+++ crates/vsock-guest/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## [0.19.66](https://github.com/vm0-ai/vm0/compare/vsock-guest-v0.19.65...vsock-guest-v0.19.66) (2026-08-07)
+
 ## [0.19.65](https://github.com/vm0-ai/vm0/compare/vsock-guest-v0.19.64...vsock-guest-v0.19.65) (2026-08-07)
 
 ## [0.19.64](https://github.com/vm0-ai/vm0/compare/vsock-guest-v0.19.63...vsock-guest-v0.19.64) (2026-08-06)